minor cleanups
This commit creates 'pup' module, so files are not under the checkout root. The
main reason for doing this was to enable coverage reports, which can be run
with `nosetests --with-coverage --cover-package=pup`.

Configuration variables have been moved to pup.utils.configuration.

A couple of unit tests were added to fact_extract.

Overall, `pup.py` is slowly being moved into task-specific files.